% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distance_hawkes.R
\name{generate_state}
\alias{generate_state}
\title{Generate the state}
\usage{
generate_state(init, n, lower, upper, sd, a = 1)
}
\arguments{
\item{init}{initial value}

\item{n}{number of states to generate}

\item{lower}{lower bound}

\item{upper}{upper bound}

\item{sd}{standard deviation of normal}

\item{a}{coefficient to apply}
}
\description{
Generate the state
}
\examples{
init <- 2.5
x <- generate_state(init, 10, 0, 3.5, 1, 0.9)
}
