% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distance_hawkes.R
\name{loss_hawkes}
\alias{loss_hawkes}
\title{Hawkes loss function}
\usage{
loss_hawkes(x, theta, time1, time2, inp)
}
\arguments{
\item{x}{state numeric}

\item{theta}{parameter numeric}

\item{time1}{first time step}

\item{time2}{second time step}

\item{inp}{inputs}
}
\description{
Hawkes loss function
}
\examples{
states <- generate_state(NULL, 2, 0, 3.5, sd = 1, 0.9)
theta <- c(0.25, 0.5)
TT <- length(states)
lambda_fun <- stepfun(seq(10, TT*10 - 10, by = 10), y = states)
kern <- function(x){return(decay_func(x, alpha = theta[1], delta = theta[2]))}
output <- sim_hawkes(lambda_fun, NULL, kern, 0, TT*10, progressBar = FALSE)
inp <- list(
  lower = 0,
  upper = 3.5,
  sd_t = 1,
  a_logit = 0.9,
  history = output,
  simulator = hawkes_simulator
)
loss_hawkes(states[2], theta, 10, 20, inp)
}
