% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SSI_initialize.R
\docType{data}
\name{Ny}
\alias{Ny}
\title{Initialise SMC}
\format{
An object of class \code{numeric} of length 1.
}
\usage{
Ny
}
\description{
Initialise SMC
}
\examples{
param <- list(alpha = 1.75, beta = 0.1, mu = -0.2, phi = 0.95, s_h = 0.6, s_v = 0.8)
tmp <- SMC.initialize(param, Nx, Ny, yobs[1])
Nx <- 1e5
}
\keyword{datasets}
