% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim_hawkes.R
\name{sim_hawkes}
\alias{sim_hawkes}
\title{Simulate inhomogenous Hawkes process}
\usage{
sim_hawkes(
  lambda_fun,
  history0 = NULL,
  kern,
  startT = max(history0),
  endT = 10,
  maxI = Inf,
  progressBar = TRUE
)
}
\arguments{
\item{lambda_fun}{function for inhomogenous poisson process}

\item{history0}{vector of historic times}

\item{kern}{function Hawkes kernal}

\item{startT}{numeric start time}

\item{endT}{numeric end time}

\item{maxI}{numeric maximum number of points to simulate}
}
\description{
Simulate inhomogenous Hawkes process
}
\examples{
lambda_levels <- c(0.5, 1, 2, 1, 3, 0.5, 1, 1, 2, 0.5, 0.5, 0.5, 0.5, 2, 2, 2, 4, 0, 1, 1, 1)
lambda_fun <- stepfun(seq(10, 200, by = 10), y = lambda_levels)
kern <- function(x){return(decay_func(x, alpha = 1, delta = 1.2))}

x <- sim_hawkes(lambda_fun, NULL, kern, 0, 200)

hist(x, breaks = 200)

plot(lambda_fun, add = TRUE, col = "red")
}
