% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Volatility_model.R
\name{stable}
\alias{stable}
\title{stable distribution}
\usage{
stable(n, alpha, beta, gamma, delta)
}
\description{
stable distribution
}
\references{
Vankov, E., & Ensor, K. B. (2014). Stochastic Volatility Filtering with Intractable Likelihoods. arXiv:1405.4323.
}
